server/requirements.txt
docling==2.15.1
sentence-transformers==4.1.0
transformers==4.51.3
torch==2.4.0
torchvision
torchaudio
flask
python-dotenv==1.0.1
docx2txt==0.8
unstructured==0.16.5
unstructured-client==0.22.0
XlsxWriter==3.2.0
python-pptx==1.0.2
openpyxl==3.1.5
tqdm
lxml==5.1.0
rank_bm25
openai==1.59.6
google-genai==1.13.0
pillow==10.4.0
ragas==0.1.15
jq
flashrank
accelerate==0.34.0
pydantic==2.10.5
pydantic_core==2.27.2
psycopg2-binary==2.9.9
deepeval
nltk==3.9.1
langchain-experimental==0.3.4